This is just an initial concept to break apart our Docker files.
This addresses #29,
and gives us a `base` image that just has the Python environment,
VCS tools,
and a couple other standard utilities.

Then it has a `full` image that adds all of the PDF/C libraries/etc.
This is where most of the time and bloat comes from,
and isn't actually required for most testing,
HTML builds,
conda builds,
and lots of other use cases.

This would allow us to interate and test base builds faster,
while keeping the environment the same in production.

NOTE: I'm mostly looking for feedback on this PR,
not specific nitpicking.