Add labels for python versions
This uses ENV to set variables that can be used for the pyenv calls, but also
allows for us to set labels of the supported Python versions, right on the
docker image.

This will be useful for setting up metadata that allows us to wipe the
virtualenv in the case of a python bugfix version number change.