Use Miniconda2 to make it backward compatible
Here is the difference,

> There are two variants of the installer: Miniconda is Python 2 based and Miniconda3 is Python 3 based. Note that the choice of which Miniconda is installed only affects the root environment. Regardless of which version of Miniconda you install, you can still install both Python 2.x and Python 3.x environments.

This one, affects us since if the user was not specifying the Python version when using `conda` we will defaulting to `python3`.

> The other difference is that the Python 3 version of Miniconda will default to Python 3 when creating new environments and building packages.

https://conda.io/miniconda.html

Considering that we are defaulting `python3` for new projects, it would be good to find a way to use `py2` only in old conda projects, though.