Re-add Python 3.5
Python 3.5 was removed at
https://github.com/rtfd/readthedocs-docker-images/pull/84/ and we
didn't have a good reason to do it.

Then, when releasing the new Docker 5.0 image we realize that we will
be breaking some projects that are pinning 3.5 in their config file:
https://github.com/rtfd/readthedocs.org/pull/5657

So, we decided to re-add it to be able to deploy this image without
breaking people's projects.