Release `ubuntu-20.04-2022.02.16`
This Docker image does not contain any change in the Dockerfile. However, as it
re-builds completely new versions of the same packages will be installed.

We are forcing this because we've hit issues with the old `ca-certificates`
version installed in our current production image. See
https://github.com/readthedocs/readthedocs.org/issues/8555

Note that we are currently executing `apt-get install ca-certificates` on each
build to upgrade this package and workaround this problem. However, this adds
extra time to _all builds_ and we want to avoid that.

Also note that this will be still required for `readthedocs/build:latest` image,
but we will hopefully use `readthedocs/build:ubuntu-20.04` as default image for
the "Clonning" step sooner than later.